create: clogger/loglevel.go
