+++ clogger/clogger.go
@@ -3,23 +3,6 @@ package clogger
 import (
 	"fmt"
 	"log"
-
-	"github.com/CTNOriginals/BitburnerGoFilesync/utils"
-)
-
-type ELogLevel utils.TBitMask
-
-const (
-	LogInfo ELogLevel = 1 << iota
-	LogDebug
-	LogWarn
-	LogError
-	LogFatal
-
-	LogNone    = 0
-	LogMessage = LogInfo | LogDebug
-	LogAlert   = LogWarn | LogError | LogFatal
-	LogAll     = (1 << iota) - 1
 )
 
 type SClog struct {
@@ -28,6 +11,7 @@ type SClog struct {
 	PrefixDate bool
 
 	LogLevel ELogLevel
+	LogState MLogLevelState
 
 	logger *log.Logger
 }
